refactor: split trade insights ai panel components
